+++ lib/vindi-rails-integrations.rb
@@ -2,6 +2,7 @@
 
 require "vindi"
 require_relative "vindi/integrations/version"
+require_relative "vindi/integrations/diagnostics"
 require_relative "vindi/integrations/railtie" if defined?(Rails)
 require_relative "vindi/integrations/concerns/synchronizable" if defined?(ActiveRecord)